+++ .golangci.yaml
@@ -1,6 +1,8 @@
 run:
   timeout: 3m
   modules-download-mode: readonly
+  build-tags:
+    - functional
 
 linters:
   enable:
